Update Quantity repair
