+++ src/main/java/com/infoshareacademy/mistrzowieVaBank/config/WebSecurityConfig.java
@@ -1,19 +1,79 @@
 package com.infoshareacademy.mistrzowieVaBank.config;
 
 
+import com.infoshareacademy.mistrzowieVaBank.MistrzowieVaBankApplication;
 import com.infoshareacademy.mistrzowieVaBank.service.UserDetailsServiceImpl;
+import org.hibernate.SessionFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.SpringApplication;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.core.env.Environment;
+import org.springframework.jdbc.datasource.DriverManagerDataSource;
+import org.springframework.orm.hibernate5.HibernateTransactionManager;
+import org.springframework.orm.hibernate5.LocalSessionFactoryBean;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
+import javax.sql.DataSource;
+import java.util.Properties;
+
 @Configuration
 @EnableWebSecurity
 public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
+
+    @Autowired
+    private Environment env;
+
+    @Bean(name = "dataSource")
+    public DataSource getDataSource() {
+        DriverManagerDataSource dataSource = new DriverManagerDataSource();
+
+        // See: application.properties
+        dataSource.setDriverClassName(env.getProperty("spring.datasource.driver-class-name"));
+        dataSource.setUrl(env.getProperty("spring.datasource.url"));
+        dataSource.setUsername(env.getProperty("spring.datasource.username"));
+        dataSource.setPassword(env.getProperty("spring.datasource.password"));
+
+        System.out.println("## getDataSource: " + dataSource);
+
+        return dataSource;
+    }
+
+    @Autowired
+    @Bean(name = "entityManagerFactory")
+    public SessionFactory getSessionFactory(DataSource dataSource) throws Exception {
+        Properties properties = new Properties();
+
+        // See: application.properties
+        properties.put("hibernate.dialect", env.getProperty("spring.jpa.properties.hibernate.dialect"));
+        properties.put("hibernate.show_sql", env.getProperty("spring.jpa.show-sql"));
+        properties.put("current_session_context_class", //
+                env.getProperty("spring.jpa.properties.hibernate.current_session_context_class"));
+
+        LocalSessionFactoryBean factoryBean = new LocalSessionFactoryBean();
+
+        // Package contain entity classes
+        factoryBean.setPackagesToScan(new String[]{""});
+        factoryBean.setDataSource(dataSource);
+        factoryBean.setHibernateProperties(properties);
+        factoryBean.afterPropertiesSet();
+        //
+        SessionFactory sf = factoryBean.getObject();
+        System.out.println("## getSessionFactory: " + sf);
+        return sf;
+    }
+
+    @Autowired
+    @Bean(name = "transactionManager")
+    public HibernateTransactionManager getTransactionManager(SessionFactory sessionFactory) {
+        HibernateTransactionManager transactionManager = new HibernateTransactionManager(sessionFactory);
+
+        return transactionManager;
+    }
     @Autowired
     UserDetailsServiceImpl userDetailsService;
 
@@ -36,5 +96,6 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
     protected void configure(HttpSecurity http) throws Exception {
 
 	}
+
 
 }
