+++ src/main/resources/templates/shoppingCart.html
@@ -35,7 +35,7 @@
                 </p>
                 
                 
-                        <input type="hidden"      
+                        <input type="hidden"
                            th:name="|cartLines[${varStatus.index}].wineInfo.id|"
                            th:value="${cartLineInfo.wineInfo.id}" />
                 <a href="" style="margin-right: 10px;">Move to Favorite</a>
